Refactor CSS imports and add styles for Menu Editor feature
