+++ frontend/src/pages/Home.jsx
@@ -24,11 +24,11 @@ function Home() {
                             <button className="nav-link nav-link-button" type="button" disabled aria-disabled="true">
                                 Log in
                             </button>
-                            <Link to="/restaurants/1/menu-editor" className="nav-link menu-editor-button">
-                                Edit menu R1
-                            </Link>
                             <Link to="/register-customer" className="nav-link signup-button">
                                 Sign up
+                            </Link>
+                            <Link to="/restaurants/1/menu-editor" className="nav-link menu-editor-button">
+                                Edit menu
                             </Link>
                         </nav>
