+++ frontend/src/pages/MenuEditor.jsx
@@ -7,7 +7,7 @@ import {
   updateDish, 
   deleteDish 
 } from "../api/authApi";
-import "../styles/MenuEditor.css";
+import "../assets/css/MenuEditor.css";
 
 const RESTAURANT_ID = 1; //WE WILL CHANGE IT WHEN WE IMPLEMENT AUTH
