--- frontend/src/styles/MenuEditor.css
@@ -1,238 +1,0 @@
-.menu-editor-page {
-  max-width: 1080px;
-  margin: 0 auto;
-  padding: 40px 20px 56px;
-}
-
-.menu-editor-panel {
-  background: linear-gradient(155deg, #ffffff 0%, #eef9ff 45%, #f5fffa 100%);
-  border: 1px solid #dbe7f5;
-  border-radius: 20px;
-  padding: 28px;
-  box-shadow: 0 10px 30px rgba(9, 20, 40, 0.08);
-}
-
-.menu-editor-header {
-  margin-bottom: 20px;
-}
-
-.menu-editor-kicker {
-  margin: 0 0 8px;
-  color: #0f766e;
-  font-weight: 700;
-  letter-spacing: 0.08em;
-  text-transform: uppercase;
-  font-size: 12px;
-}
-
-.menu-editor-header h1 {
-  margin: 0;
-  font-size: clamp(30px, 4vw, 42px);
-}
-
-.menu-editor-header p {
-  margin: 8px 0 0;
-  color: #5d6b87;
-}
-
-.menu-editor-grid {
-  display: grid;
-  gap: 20px;
-  grid-template-columns: 1.1fr 1fr;
-}
-
-.menu-editor-form,
-.menu-editor-list {
-  background: #ffffff;
-  border: 1px solid #dbe7f5;
-  border-radius: 16px;
-  padding: 18px;
-}
-
-.menu-editor-form h2,
-.menu-editor-list h2 {
-  margin-top: 0;
-}
-
-.menu-editor-form label {
-  display: block;
-  margin-bottom: 6px;
-  font-weight: 700;
-  color: #25314a;
-}
-
-.menu-editor-form input,
-.menu-editor-form textarea {
-  width: 100%;
-  margin-bottom: 12px;
-  border: 1px solid #c8d9ef;
-  border-radius: 10px;
-  padding: 10px 12px;
-  font: inherit;
-  background: #fcfdff;
-}
-
-.menu-editor-form input:focus,
-.menu-editor-form textarea:focus {
-  outline: 3px solid #99f6e4;
-  outline-offset: 1px;
-}
-
-.menu-editor-form fieldset {
-  border: 1px solid #cfe0f8;
-  border-radius: 12px;
-  margin: 0;
-  margin-bottom: 14px;
-  padding: 12px;
-}
-
-.menu-editor-form legend {
-  font-weight: 700;
-  color: #25314a;
-  padding: 0 6px;
-}
-
-.allergen-list {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 8px;
-}
-
-.allergen-chip {
-  display: inline-flex;
-  align-items: center;
-  gap: 6px;
-  padding: 6px 10px;
-  border-radius: 999px;
-  background: #f2f9ff;
-  border: 1px solid #d0e2f7;
-  color: #25314a;
-  font-size: 14px;
-  cursor: pointer;
-}
-
-.allergen-chip input {
-  margin: 0;
-}
-
-.menu-editor-actions {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 10px;
-}
-
-.menu-editor-message {
-  margin: 12px 0 0;
-  color: #0f766e;
-  font-weight: 700;
-}
-
-.menu-editor-list ul {
-  margin: 0;
-  padding: 0;
-  list-style: none;
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-}
-
-.dish-card {
-  border: 1px solid #dbe7f5;
-  border-radius: 12px;
-  padding: 14px;
-  display: flex;
-  justify-content: space-between;
-  gap: 10px;
-  align-items: flex-start;
-  background: #fcfeff;
-}
-
-.dish-card h3 {
-  margin: 0 0 6px;
-}
-
-.dish-description,
-.dish-price,
-.dish-allergens {
-  margin: 0 0 6px;
-  color: #31415f;
-}
-
-.dish-price {
-  font-weight: 700;
-}
-
-.dish-actions {
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-}
-
-.btn-danger {
-  background: #dc2626;
-  color: #fff;
-  box-shadow: 0 6px 16px rgba(220, 38, 38, 0.24);
-}
-
-.menu-editor-form input:disabled,
-.menu-editor-form textarea:disabled {
-  background: #f0f4f9;
-  color: #9ca3af;
-  cursor: not-allowed;
-}
-
-.menu-editor-form button[disabled],
-.menu-editor-form button:disabled {
-  opacity: 0.6;
-  cursor: not-allowed;
-}
-
-.btn:disabled {
-  opacity: 0.6;
-  cursor: not-allowed;
-}
-
-.allergen-chip input:disabled {
-  cursor: not-allowed;
-  opacity: 0.6;
-}
-
-.allergen-loading,
-.allergen-error,
-.allergen-empty,
-.loading-message,
-.error-message,
-.empty-message {
-  color: #5d6b87;
-  font-size: 14px;
-  margin: 8px 0;
-}
-
-.allergen-error,
-.error-message {
-  color: #dc2626;
-  font-weight: 600;
-}
-
-.menu-editor-footer-link {
-  margin-top: 20px;
-}
-
-@media (max-width: 900px) {
-  .menu-editor-panel {
-    padding: 22px;
-  }
-
-  .menu-editor-grid {
-    grid-template-columns: 1fr;
-  }
-
-  .dish-card {
-    flex-direction: column;
-  }
-
-  .dish-actions {
-    width: 100%;
-    flex-direction: row;
-  }
-}
